backend/requirements.txt
boto3==1.34.28
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
